first implementation of the LinstorClient class
This is an abstraction of all client call to the linstor api.
Currently node and netinterface commands are implemented.
After tests and some research the socket communication was written by hand
in a non blocking behavior using select and RLocks.
Prior to that an asynchat and twisted version was tested, but
both implementations had their downsides:
  asynchat: is not thread-safe
  twisted: hard to integrate in other python applications if they aren't
           prepared for event driven style