linstorapi: add resource_delete_if_diskless
While this looks a bit esoteric first, this is an actual, not that
uncommon use case for plugins (e.g., xen). In such scenarios a resource
is usually created n-times redundant, and then it is "attached". In that
case a) the assignment already exists or b) is created as diskless
assignment. On "detach", we only want to remove the resource from the
node if it was diskless, but not if it is part of the static redundancy.

Adding an optional parameter e.g., if_diskless to resource_delete() felt
too messy.