linstorapi: add resource_create_and_auto_place
This function/method is pretty common for plugins. These usually have a
"create" call, where the storage is created N-times redundant. This
method provides a convenient way to do that.

Later these plugins get an "open" call on some node, then diskless
assignments are created as needed. The API already has convenient
methods for that.