fix a possible too early exit from wait loop
Since we now use a callback function to decide if we keep waiting
for api answers, and clear the _replies map entry afterward.
The check if the _replies map entry has 0 entries is wrong,
as this would lead to leaving the loop if there are currently
no entries in the map, but there could still be incoming.

Answers for a message id are terminated with the COMPLETE message.