linstorapi: add resource_dfn_props_list
We already have a convenient way to set resource definition properties,
but we lack a more convenient way to retrieve them. This implements
property retrieval filtered by a given namespace.

Initially this was triggered by Xen, as volume plugins have to handle
meta data (e.g., description). Aux-properties are a perfect fit for
that.