linstorapi: fix concurrency bug run()/disconnect()
While working on Xen I saw multiple cases where the API raised
exceptions. This was cased because run() and disconnect() have not been
sufficiently synchronized.

In general I would propose a solution where disconnect() signals to
run() that its job is done. Then run() signals back that it has finished
and then disconnect() returns.

Unfortunately run() relies on checking for _socket is None, and I did
not want to break that. So the simple solution is to still let run()
fail/check _socket that is still closed by disconnect() but simply have
an extra lock that synchronizes run() return and the eventual
disconnect() return.