introduce a higher-level API
While the current API is in a technically good shape, there is need for
an higher level abstraction:

- The low-level API shows/exposes too many implementation details. One
has to be (very) familiar with the underlying protobuf structs to get
something done. In general a user facing API should not expose
implementation details.
- One needs to stack multiple calls/error handling to get common things
done (e.g., delete a diskless assignment, but only if it is diskless).
- Some common things are completely hidden (e.g., setting properties like
'allow-two-primaries').

This is an approach to change that and add this higher-level of
abstraction. For examples, please see/generate the documentation in the
doc/ directory. Note that this is not a complete abstraction of all
objects in LINSTOR as it misses for example higher-level abstractions
for nodes or storage pools. So far the API only contains objects
necessary for plugin devs. If there is need for additional higher-level
objects, we add them.