config: add linstor.Config class
This is a simple helper class that exposes three functions:

- read_config: uses SafeConfigParser() to parse a config file and
returns its sections in a dictionary. Can be used directly, is mainly
used by other functions in this class

- get_section: returns a section (e.g., 'global') from a config file. If
no config file is specified, this function searches the default LINSTOR
client configuration files.

- get_controllers: This is a opinionated function that returns output
that can be used with MultiLinstor. If the environment variable
'LS_CONTROLLERS' is set, its value will be used. If not, the usual
config files are searched (call to get_section()), and finally a
fallback is used (by default linstor://localhost).