high-level API: add ResourceGroup object
This new class represents a LINSTOR resource group and makes it
accessible in the high level API. It allows what one would expect,
including creating, modifying, deleting resource groups and creating new
Resource objects from a ResourceGroup.

Important: This modifies the semantic of Resource.from_resource_group().
This allows it to follow the semantic used in the Resource object (as
well as ResourceGroup). We don't have that many users already using
resource groups. Overall this makes it a lot more consistent with the
rest of the HL API.