high-level API: expose query_max_volume_size in RG
This is a simple wrapper for the low level query_max_volume_size()
method. For the high level API we could beef it up by for example having
a parameter to only return the best (== max size) candidate. But for now
just expose it as it looks pretty useful as is.