Add option to re-use a client object in high-level API
The high-level API had limited customization options for the LINSTOR client
it was using internally. This meant for example: no option to configure HTTPS
certificates or timeouts. To remedy this, allow using a pre-configured client.

This client is passed in on object construction and reused instead of
constructing a new client on every use. One change that is required for this
to work is for the context manager on the client to count the number of
"active" users and only disconnecting the socket once the count reaches zero.
Otherwise the client is closed prematurely once the innermost "with" block
is exited.